web: unify procedure call interface and seperation of concerns
Seperate out the web from individual procedures, which take an input
json as arguments, and return a json or an error.

The procedure name will be represented as the URL path without the
first forward slash, such as `/analyze` or `/whatever`.  If the
procedure is not found, a 404 will be returned.

If it returns an error, the Error() will be called and wrapped in
a `{"error": }` object.  And the status will be 500.

If it returns a result, which will be of type `*bytes.Buffer`, and
will be consist of a json string, and the result will be wraped in
a `{"data": }` object.  And the status will be 200.